move education utils
