+++ apps/web/src/utils/educationUtils.tsx
